% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stream.R
\name{geom_stream}
\alias{geom_stream}
\title{geom_stream

A streamgraph geom/stat for ggplot2}
\usage{
geom_stream(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  na.rm = T,
  bw = 0.75,
  n_grid = 3000,
  ...
)
}
\arguments{
\item{mapping}{provide you own mapping. both x and y need to be numeric.}

\item{data}{provide you own data}

\item{geom}{change geom}

\item{position}{change position}

\item{show.legend}{show legend in plot}

\item{inherit.aes}{should the geom inherits aestethics}

\item{na.rm}{remove missing values}

\item{bw}{bandwidth of kernal density estimation}

\item{n_grid}{number of x points that should be calculated. The higher the more smooth plot.}

\item{...}{other arguments to be passed to the geom}
}
\value{
a data frame
}
\description{
geom_stream

A streamgraph geom/stat for ggplot2
}
