% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stream.R
\name{make_connect_dots}
\alias{make_connect_dots}
\title{make_connect_dots

Returns n number of points from data that perfectly fits data.}
\usage{
make_connect_dots(.df, n_grid = n_grid, min_x, max_x)
}
\arguments{
\item{.df}{a data frame that must contain x and y}

\item{n_grid}{number of x points that should be calculated. The higher the more smooth plot.}

\item{min_x}{minimum x value of all groups}

\item{max_x}{maximum x value of all groups}
}
\value{
a data frame
}
\description{
make_connect_dots

Returns n number of points from data that perfectly fits data.
}
