% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stream.R
\name{stack_densities}
\alias{stack_densities}
\title{stack_densities

Takes density lines of equal x range and stack them on top of each other symmetrically aournd zero.}
\usage{
stack_densities(data, bw = bw, n_grid = n_grid)
}
\arguments{
\item{data}{a data frame}

\item{bw}{bandwidth of kernal density}

\item{n_grid}{number of x points that should be calculated. The higher the more smooth plot.}
}
\value{
a data frame
}
\description{
stack_densities

Takes density lines of equal x range and stack them on top of each other symmetrically aournd zero.
}
