% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stream.R
\docType{data}
\name{StatStreamDensity}
\alias{StatStreamDensity}
\title{StatStreamDensity

stat to compute `geom_stream`}
\format{
An object of class \code{StatStreamDensity} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
StatStreamDensity
}
\arguments{
\item{mapping}{provide you own mapping. both x and y need to be numeric.}

\item{data}{provide you own data}

\item{geom}{change geom}

\item{position}{change position}

\item{na.rm}{remove missing values}

\item{show.legend}{show legend in plot}

\item{bw}{bandwidth of kernal density estimation}

\item{n_grid}{number of x points that should be calculated. The higher the more smooth plot.}

\item{inherit.aes}{should the geom inherits aestethics}

\item{...}{other arguments to be passed to the geom}
}
\value{
a data frame
}
\description{
StatStreamDensity

stat to compute `geom_stream`
}
\keyword{datasets}
