% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stream.R
\name{stack_densities}
\alias{stack_densities}
\title{stack_densities

Takes density lines of equal x range and stack them on top of each other symmetrically aournd zero.}
\usage{
stack_densities(
  data,
  bw = bw,
  n_grid = n_grid,
  center_fun = center_fun,
  method = method
)
}
\arguments{
\item{data}{a data frame}

\item{bw}{bandwidth of kernal density}

\item{n_grid}{number of x points that should be calculated. The higher the more smooth plot.}

\item{center_fun}{a function that returns the y center for each possible x in range of x.}

\item{method}{which estimation should be applied. Default is LOESS.}
}
\value{
a data frame
}
\description{
stack_densities

Takes density lines of equal x range and stack them on top of each other symmetrically aournd zero.
}
